General Improvement.

+++ app/Http/Controllers/Tip/TipDelete.php
@@ -18,11 +18,20 @@ class TipDelete extends TipController
 	*/
     public function destroy($id)
     {
+		try{
         $tip = Tip::find($id);
+		if($tip ==null){
+			 return json_encode(array('message'=>'Value not found!'));
+		}
         $tip->delete();	
 		$answer=array('message'=>"Tip deletet sucessfully!");
         return json_encode($answer);
-    }
+		}
+		catch(Exception $e) {
+		$answer = array('status' => 'error','message' => $e->getMessage());
+		return json_encode($answer); 
+		}
+	}
 }
 
 ?>
